requirements.txt
--extra-index-url https://pypi.celonis.cloud/
pandas
pycelonis>=2.0.0
openpyxl
python-dotenv
